Change the search field's value when the map selection is changed.
Clear the map selection when the search field is cleared.

Skip making calls to the Geocoding API when the search suggestion menu is closed.

Find map regions by looking them up in the GeoJSON instead of faking clicks on the map or calling querySourceFeatures(), as neither of those can find features not currently in view.

Position map tooltips using centerOfMass() instead of center(). center() doesn't care what the shape is. It averages all of the points, so a polygon with a more complicated bottom edge would have a lower center. centerOfMass() is actually the center of the shape.

Rewrite Header as a functional component.