Plug leak of MapPopupContent with a MUI Portal
The use of `ReactDOM.render` was leading to the creation and subsequent leak of a new MapPopupContent for every map feature moused over. This commit uses a single, persistent MapPopupContent wrapped in a MUI Portal to ensure there's only ever one created.

As a bonus, this also puts MapPopupContent back into the normal React render hierarchy, giving it access to the same ThemeProvider as everything else and removing the need for reiterating styles in MapPopupContent.

This also makes sure to not reset the hoverPopup's link to the MapPopupContent when it is already visible, making it possible for MapPopupContent's content to animate transitions between hovered regions, if that ends up being something desirable.